correções

+++ commands/database/rstatus.js
@@ -1,5 +1,6 @@
 const Discord = require("discord.js")
 const db = require("quick.db")
+const config = require("../../config.json")
 
 module.exports = {
   name: "rstatus",
@@ -9,8 +10,9 @@ module.exports = {
   description: "Remove o status do bot.",
   run: (client, message, args) => {
     
-    if (message.member.user.id == '459559578648969227' ) {
-    }else{ return message.reply(`Somente o meu criador pode usar esse comando`) }
+    if(message.author.id != config.donoID) {
+      return message.channel.send("Esse comando é especifico para o meu criador.")
+    }
      
     db.delete(`status_`, )
